Change input to float

+++ scripts/calculator.py
@@ -1,6 +1,6 @@
 print("Welcome to the super calculator yeahhh!\n")
-a = int(input("Enter the first integer: "))
-b = int(input("Enter the second integer: "))
+a = float(input("Enter the first integer: "))
+b = float(input("Enter the second integer: "))
 
 print("The sum of both numbers is " + str(a+b))
